(refactor) organized assets folder, imported json animations
