+++ thesis-client/src/assets/animation/loader.json
@@ -1,0 +1,1 @@
+{"v":"4.5.7","fr":30,"ip":0,"op":97,"w":512,"h":512,"ddd":0,"assets":[{"id":"comp_0","layers":[{"ddd":0,"ind":0,"ty":4,"nm":"Layer 2/loader Outlines","ks":{"o":{"k":100},"r":{"k":0},"p":{"k":[{"i":{"x":0.667,"y":1},"o":{"x":0.167,"y":0.167},"n":"0p667_1_0p167_0p167","t":15,"s":[400,337,0],"e":[396,335.04,0],"to":[0,-0.57204169034958,0],"ti":[2.25,3.28968811035156,0]},{"i":{"x":0.833,"y":0.833},"o":{"x":0.31,"y":0},"n":"0p833_0p833_0p31_0","t":26,"s":[396,335.04,0],"e":[400,230,0],"to":[-1.89180493354797,-2.76597690582275,0],"ti":[32.75,5.75,0]},{"i":{"x":0.658,"y":0.366},"o":{"x":0.322,"y":0},"n":"0p658_0p366_0p322_0","t":37,"s":[400,230,0],"e":[397.463,229.697,0],"to":[-0.8968859910965,-0.13742607831955,0],"ti":[0.79534405469894,0.06496899574995,0]},{"i":{"x":0.749,"y":0.753},"o":{"x":0.356,"y":0.179},"n":"0p749_0p753_0p356_0p179","t":38,"s":[397.463,229.697,0],"e":[400,265,0],"to":[-26.6949577331543,-2.18062162399292,0],"ti":[0,0,0]},{"i":{"x":0.811,"y":0},"o":{"x":0.167,"y":0},"n":"0p811_0_0p167_0","t":43,"s":[400,265,0],"e":[470.35,266.825,0],"to":[0,0,0],"ti":[0,0,0]},{"t":83.0000033806593}]},"a":{"k":[106,105.5,0]},"s":{"k":[{"i":{"x":[0.207,0.207,0.667],"y":[0.954,0.954,0.667]},"o":{"x":[0.167,0.167,0.167],"y":[0,0,0.167]},"n":["0p207_0p954_0p167_0","0p207_0p954_0p167_0","0p667_0p667_0p167_0p167"],"t":0,"s":[0,0,100],"e":[33,33,100]},{"t":15.0000006109625}]}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ks":{"k":{"i":[[0,57.99],[57.99,0],[0,-57.99],[-57.99,0]],"o":[[0,-57.99],[-57.99,0],[0,57.99],[57.99,0]],"v":[[105,0],[0,-105],[-105,0],[0,105]],"c":true}},"nm":"Path 1","mn":"ADBE Vector Shape - Group"},{"ty":"fl","fillEnabled":true,"c":{"k":[0.08,0.08,0.08,1]},"o":{"k":100},"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill"},{"ty":"tr","p":{"k":[106,105.5],"ix":2},"a":{"k":[0,0],"ix":1},"s":{"k":[100,100],"ix":3},"r":{"k":0,"ix":6},"o":{"k":100,"ix":7},"sk":{"k":0,"ix":4},"sa":{"k":0,"ix":5},"nm":"Transform"}],"nm":"Group 1","np":2,"mn":"ADBE Vector Group"}],"ip":0,"op":43.0000017514259,"st":-3.00000012219251,"bm":0,"sr":1},{"ddd":0,"ind":1,"ty":4,"nm":"Layer 1/loader Outlines","ks":{"o":{"k":100},"r":{"k":[{"i":{"x":[0.428],"y":[1]},"o":{"x":[0.167],"y":[0.167]},"n":["0p428_1_0p167_0p167"],"t":43,"s":[0],"e":[90]},{"i":{"x":[0.517],"y":[1]},"o":{"x":[0.359],"y":[0]},"n":["0p517_1_0p359_0"],"t":58,"s":[90],"e":[480]},{"i":{"x":[0.128],"y":[0.158]},"o":{"x":[0.249],"y":[0]},"n":["0p128_0p158_0p249_0"],"t":83,"s":[480],"e":[500]},{"i":{"x":[0.762],"y":[-9.662]},"o":{"x":[0.958],"y":[1.387]},"n":["0p762_-9p662_0p958_1p387"],"t":87,"s":[500],"e":[510]},{"i":{"x":[0.833],"y":[1]},"o":{"x":[0],"y":[0]},"n":["0p833_1_0_0"],"t":90,"s":[510],"e":[900]},{"t":106.000004317469}]},"p":{"k":[400,337.175,0]},"a":{"k":[105.5,105.5,0]},"s":{"k":[100,100,100]}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ks":{"k":{"i":[[0,19.33],[19.33,0],[0,-19.33],[-19.33,0]],"o":[[0,-19.33],[-19.33,0],[0,19.33],[19.33,0]],"v":[[35,0],[0,-35],[-35,0],[0,35]],"c":true}},"nm":"Path 1","mn":"ADBE Vector Shape - Group"},{"ty":"fl","fillEnabled":true,"c":{"k":[0.08,0.08,0.08,1]},"o":{"k":100},"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill"},{"ty":"tr","p":{"k":[105.5,175.5],"ix":2},"a":{"k":[0,0],"ix":1},"s":{"k":[{"i":{"x":[0.667,0.667],"y":[1,1]},"o":{"x":[0.333,0.333],"y":[0,0]},"n":["0p667_1_0p333_0","0p667_1_0p333_0"],"t":45,"s":[0,0],"e":[100,100]},{"i":{"x":[0.833,0.833],"y":[0.833,0.833]},"o":{"x":[0.167,0.167],"y":[0.167,0.167]},"n":["0p833_0p833_0p167_0p167","0p833_0p833_0p167_0p167"],"t":51,"s":[100,100],"e":[100,100]},{"i":{"x":[0.833,0.833],"y":[1,1]},"o":{"x":[0.167,0.167],"y":[0,0]},"n":["0p833_1_0p167_0","0p833_1_0p167_0"],"t":98,"s":[100,100],"e":[0,0]},{"t":104.000004236007}],"ix":3},"r":{"k":0,"ix":6},"o":{"k":100,"ix":7},"sk":{"k":0,"ix":4},"sa":{"k":0,"ix":5},"nm":"Transform"}],"nm":"Group 1","np":2,"mn":"ADBE Vector Group"},{"ty":"gr","it":[{"ind":0,"ty":"sh","ks":{"k":{"i":[[0,19.33],[19.33,0],[0,-19.33],[-19.33,0]],"o":[[0,-19.33],[-19.33,0],[0,19.33],[19.33,0]],"v":[[35,0],[0,-35],[-35,0],[0,35]],"c":true}},"nm":"Path 1","mn":"ADBE Vector Shape - Group"},{"ty":"fl","fillEnabled":true,"c":{"k":[0.08,0.08,0.08,1]},"o":{"k":100},"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill"},{"ty":"tr","p":{"k":[105.5,35.5],"ix":2},"a":{"k":[0,0],"ix":1},"s":{"k":[{"i":{"x":[0.833,0.833],"y":[0.833,0.833]},"o":{"x":[0.167,0.167],"y":[0.167,0.167]},"n":["0p833_0p833_0p167_0p167","0p833_0p833_0p167_0p167"],"t":43,"s":[100,100],"e":[100,100]},{"i":{"x":[0.833,0.833],"y":[1,1]},"o":{"x":[0.167,0.167],"y":[0,0]},"n":["0p833_1_0p167_0","0p833_1_0p167_0"],"t":94,"s":[100,100],"e":[0,0]},{"t":100.000004073084}],"ix":3},"r":{"k":0,"ix":6},"o":{"k":100,"ix":7},"sk":{"k":0,"ix":4},"sa":{"k":0,"ix":5},"nm":"Transform"}],"nm":"Group 2","np":2,"mn":"ADBE Vector Group"},{"ty":"gr","it":[{"ind":0,"ty":"sh","ks":{"k":{"i":[[0,19.33],[19.33,0],[0,-19.33],[-19.33,0]],"o":[[0,-19.33],[-19.33,0],[0,19.33],[19.33,0]],"v":[[35,0],[0,-35],[-35,0],[0,35]],"c":true}},"nm":"Path 1","mn":"ADBE Vector Shape - Group"},{"ty":"fl","fillEnabled":true,"c":{"k":[0.08,0.08,0.08,1]},"o":{"k":100},"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill"},{"ty":"tr","p":{"k":[35.5,105.5],"ix":2},"a":{"k":[0,0],"ix":1},"s":{"k":[{"i":{"x":[0.667,0.667],"y":[1,1]},"o":{"x":[0.333,0.333],"y":[0,0]},"n":["0p667_1_0p333_0","0p667_1_0p333_0"],"t":47,"s":[0,0],"e":[100,100]},{"i":{"x":[0.833,0.833],"y":[0.833,0.833]},"o":{"x":[0.167,0.167],"y":[0.167,0.167]},"n":["0p833_0p833_0p167_0p167","0p833_0p833_0p167_0p167"],"t":53,"s":[100,100],"e":[100,100]},{"i":{"x":[0.833,0.833],"y":[1,1]},"o":{"x":[0.167,0.167],"y":[0,0]},"n":["0p833_1_0p167_0","0p833_1_0p167_0"],"t":100,"s":[100,100],"e":[0,0]},{"t":106.000004317469}],"ix":3},"r":{"k":0,"ix":6},"o":{"k":100,"ix":7},"sk":{"k":0,"ix":4},"sa":{"k":0,"ix":5},"nm":"Transform"}],"nm":"Group 3","np":2,"mn":"ADBE Vector Group"},{"ty":"gr","it":[{"ind":0,"ty":"sh","ks":{"k":{"i":[[0,19.33],[19.33,0],[0,-19.33],[-19.33,0]],"o":[[0,-19.33],[-19.33,0],[0,19.33],[19.33,0]],"v":[[35,0],[0,-35],[-35,0],[0,35]],"c":true}},"nm":"Path 1","mn":"ADBE Vector Shape - Group"},{"ty":"fl","fillEnabled":true,"c":{"k":[0.08,0.08,0.08,1]},"o":{"k":100},"nm":"Fill 1","mn":"ADBE Vector Graphic - Fill"},{"ty":"tr","p":{"k":[175.5,105.5],"ix":2},"a":{"k":[0,0],"ix":1},"s":{"k":[{"i":{"x":[0.667,0.667],"y":[1,1]},"o":{"x":[0.333,0.333],"y":[0,0]},"n":["0p667_1_0p333_0","0p667_1_0p333_0"],"t":43,"s":[0,0],"e":[100,100]},{"i":{"x":[0.833,0.833],"y":[0.833,0.833]},"o":{"x":[0.167,0.167],"y":[0.167,0.167]},"n":["0p833_0p833_0p167_0p167","0p833_0p833_0p167_0p167"],"t":49,"s":[100,100],"e":[100,100]},{"i":{"x":[0.833,0.833],"y":[1,1]},"o":{"x":[0.167,0.167],"y":[0,0]},"n":["0p833_1_0p167_0","0p833_1_0p167_0"],"t":96,"s":[100,100],"e":[0,0]},{"t":102.000004154545}],"ix":3},"r":{"k":0,"ix":6},"o":{"k":100,"ix":7},"sk":{"k":0,"ix":4},"sa":{"k":0,"ix":5},"nm":"Transform"}],"nm":"Group 4","np":2,"mn":"ADBE Vector Group"}],"ip":43.0000017514259,"op":117.000004765508,"st":-1.00000004073083,"bm":0,"sr":1}]}],"layers":[{"ddd":0,"ind":0,"ty":0,"nm":"loader-source","refId":"comp_0","ks":{"o":{"k":100},"r":{"k":0},"p":{"k":[400,263.5,0]},"a":{"k":[400,300,0]},"s":{"k":[100,100,100]}},"ao":0,"w":800,"h":600,"ip":0,"op":117.000004765508,"st":0,"bm":0,"sr":1},{"ddd":0,"ind":1,"ty":1,"nm":"Medium Red Solid 1","ks":{"o":{"k":100},"r":{"k":0},"p":{"k":[400,300,0]},"a":{"k":[400,300,0]},"s":{"k":[100,100,100]}},"ao":0,"sw":800,"sh":600,"sc":"#ea6262","ip":0,"op":117.000004765508,"st":0,"bm":0,"sr":1}],"v":"4.5.3","ddd":0,"ip":0,"op":117.000004765508,"fr":29.9700012207031,"w":800,"h":600}
