+++ thesis-client/src/assets/icons/bikeIcon.png
