+++ thesis-client/src/assets/icons/homeIcon.png
