+++ thesis-client/src/assets/icons/mapIcon.png
