+++ thesis-client/src/assets/images/bikemap_logo.png
