+++ thesis-client/src/components/login-component.js
@@ -1,27 +1,29 @@
 import React from 'react';
 import { StyleSheet, Image, View, Button } from 'react-native';
-import { LinearGradient } from 'expo';
+import { LinearGradient, DangerZone } from 'expo';
 import PropTypes from 'prop-types';
+
+const { Lottie } = DangerZone;
 
 const styles = StyleSheet.create({
   box: {
     flex: 10,
     alignItems: "center",
-    justifyContent: "center"
+    justifyContent: "center",
   },
   boxImage: {
     flex: 5.5,
     alignItems: "center",
-    justifyContent: "flex-end"
+    justifyContent: "flex-end",
   },
   boxButton: {
     flex: 4.5,
-    justifyContent: "center"
+    justifyContent: "center",
   },
   image: {
     flex: 0.5,
-    resizeMode: "contain"
-  }
+    resizeMode: "contain",
+  },
 });
 
 const LoginView = ({ _handlePressAsync, disableButton }) => (
@@ -39,7 +41,7 @@ const LoginView = ({ _handlePressAsync, disableButton }) => (
     <View style={[styles.boxImage]}>
       <Image
         style={styles.image}
-        source={require('../assets/bikemap_logo.png')}
+        source={require('../assets/images/bikemap_logo.png')}
       />
     </View>
     <View style={[styles.boxButton]}>
