+++ thesis-client/src/sagas/sagas.js
@@ -56,7 +56,7 @@ const getTripsAsync = function* () {
     const tripsRequest = yield call(dbSecureGET, 'route');
     yield put({ type: 'GET_TRIPS_SUCCESS', payload: tripsRequest });
   } catch (error) {
-    console.log(error);
+    console.log('async', JSON.stringify(error));
   }
 };
