+++ thesis-client/src/screens/homescreen.js
@@ -6,7 +6,7 @@ import Expo from 'expo';
 import selectTrip from '../actions/activeTrip-action';
 import getTrips from '../actions/getTrip-action';
 import { getUserLocation } from '../actions/getUserLocation-action';
-import icon from '../assets/bikeIcon.png';
+import icon from '../assets/icons/bikeIcon.png';
 import { STATUS_BAR_HEIGHT } from '../constants';
 import Trip from '../components/trip-component';
